v0.3 완료

+++ src/main/java/com/dace/dmgr/DMGR.java
@@ -1,16 +1,18 @@
 package com.dace.dmgr;
 
 import com.dace.dmgr.combat.event.CombatEventManager;
-import com.dace.dmgr.config.GeneralConfig;
 import com.dace.dmgr.event.MainEventManager;
-import com.dace.dmgr.lobby.Lobby;
 import com.dace.dmgr.lobby.User;
-import com.dace.dmgr.system.command.*;
+import com.dace.dmgr.system.EntityInfoRegistry;
+import com.dace.dmgr.system.SystemPrefix;
+import com.dace.dmgr.system.command.LobbyCommand;
+import com.dace.dmgr.system.command.MenuCommand;
+import com.dace.dmgr.system.command.PlayerOptionCommand;
+import com.dace.dmgr.system.command.test.DummyCommand;
+import com.dace.dmgr.system.command.test.SelectCharCommand;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
-
-import static com.dace.dmgr.system.HashMapList.userMap;
 
 /**
  * 플러그인 메인 클래스.
@@ -30,21 +32,19 @@ public class DMGR extends JavaPlugin {
      */
     @Override
     public void onEnable() {
-        getLogger().info(PREFIX.LOG + "플러그인 활성화 완료");
-        GeneralConfig.init();
+        getLogger().info(SystemPrefix.LOG + "플러그인 활성화 완료");
         MainEventManager.init();
         CombatEventManager.init();
-        registerCommand();
+        registerCommands();
+        registerTestCommands();
 
         Bukkit.getOnlinePlayers().forEach((Player player) -> {
             User user = new User(player);
-            userMap.put(player, user);
-            Lobby.lobbyTick(player);
-            getServer().broadcastMessage(PREFIX.CHAT + "플레이어 할당 : §e§n" + player.getName());
+            user.init();
+            getServer().broadcastMessage(SystemPrefix.CHAT + "플레이어 등록 : §e§n" + player.getName());
         });
-        Bukkit.getOnlinePlayers().forEach((Player player) -> {
-            player.sendMessage(PREFIX.CHAT + "시스템 재부팅 완료");
-        });
+        Bukkit.getOnlinePlayers().forEach((Player player) ->
+                player.sendMessage(SystemPrefix.CHAT + "시스템 재부팅 완료"));
     }
 
     /**
@@ -52,34 +52,30 @@ public class DMGR extends JavaPlugin {
      */
     @Override
     public void onDisable() {
-        getLogger().info(PREFIX.LOG + "플러그인 비활성화 완료");
+        getLogger().info(SystemPrefix.LOG + "플러그인 비활성화 완료");
 
         Bukkit.getOnlinePlayers().forEach((Player player) -> {
-            User user = new User(player);
-            user.getLobbySidebar().delete();
-            player.sendMessage(PREFIX.CHAT + "시스템 재부팅 중...");
+            User user = EntityInfoRegistry.getUser(player);
+            if (user != null)
+                user.getLobbySidebar().delete();
+            player.sendMessage(SystemPrefix.CHAT + "시스템 재부팅 중...");
         });
     }
 
     /**
      * 모든 명령어를 등록한다.
      */
-    private void registerCommand() {
+    private void registerCommands() {
         getCommand("스폰").setExecutor(new LobbyCommand());
         getCommand("메뉴").setExecutor(new MenuCommand());
         getCommand("설정").setExecutor(new PlayerOptionCommand());
-        getCommand("선택").setExecutor(new SelectCharCommand());
     }
 
     /**
-     * 시스템 메시지의 접두사.
+     * 모든 테스트용 명령어를 등록한다.
      */
-    public static class PREFIX {
-        /** 시스템 로그 */
-        public final static String LOG = "[ DMGR-Core ] ";
-        /** 일반 */
-        public final static String CHAT = "§3§l[ §b§lDMGR §3§l] §f";
-        /** 경고 */
-        public final static String CHAT_WARN = "§3§l[ §b§lDMGR §3§l] §c";
+    private void registerTestCommands() {
+        getCommand("선택").setExecutor(new SelectCharCommand());
+        getCommand("소환").setExecutor(new DummyCommand());
     }
 }
